requirements.txt
streamlit==1.35.0
pandas==2.2.2
python-dotenv==1.0.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-groq==0.1.4
langchain-huggingface==0.0.8
faiss-cpu==1.7.4
sentence-transformers==2.7.0
torch==2.2.2
numpy==1.26.4
tqdm==4.66.4
